bootstrap: Add ACME standalone mode switch
The default mode for bootstrap expects a working webserver configured to
serve the webroot directory. This is because most DDV server nodes serve
services over the web. Certain services, however, do not serve over the
web, so there is no need to install a web server. Add standalone mode
for these scenarios.